refactor(modal): update modal to accept ReactNode description
