+++ apps/dashboard/src/utils/constants.ts
@@ -12,8 +12,3 @@ export const EXCLUDED_EDITOR_TYPES: string[] = [
   StepTypeEnum.TRIGGER,
   StepTypeEnum.CUSTOM,
 ];
-
-export const PAUSE_MODAL_TITLE = 'Proceeding will pause the workflow';
-// convert it to accept dynamic workflow name
-export const PAUSE_MODAL_DESCRIPTION = (workflowName: string) =>
-  `The ${workflowName} cannot be triggered if paused, please confirm to proceed.`;
